update schema
Updated the schema to include an owner_id for the tales table, so that
an owner can be assigned.  Added a complete column so that the story
will no longer be able to be edited.  And added a counter to the
user_tales so that you have to wait a set number of turns before they
can add the story again.