debug private users update content
there was an error occurring when trying to update a private tale from
a user on the front end.  The bug was because I was inputting the user
id as a string, so I converted instances of params[:id] to integers,
this method doesn’t cause an error if you call it on an integer.